+++ mitmproxy/proxy2/layers/__init__.py
@@ -1,7 +1,8 @@
 from . import modes
 from .glue import GlueLayer
 from .http import HTTPLayer
-from .http2 import ClientHTTP2Layer
+from .http.http1 import ClientHTTP1Layer, ServerHTTP1Layer
+from .http.http2 import ClientHTTP2Layer, ServerHTTP2Layer
 from .tcp import TCPLayer
 from .tls import ClientTLSLayer, ServerTLSLayer
 from .websocket import WebsocketLayer
@@ -9,8 +10,9 @@ from .websocket import WebsocketLayer
 __all__ = [
     "modes",
     "GlueLayer",
-    "HTTPLayer",
-    "ClientHTTP2Layer",
+    "HTTPLayer", # TODO remove this and replace with ClientHTTP1Layer
+    "ClientHTTP1Layer", "ServerHTTP1Layer",
+    "ClientHTTP2Layer", "ServerHTTP2Layer",
     "TCPLayer",
     "ClientTLSLayer", "ServerTLSLayer",
     "WebsocketLayer",
